backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Pydantic & Settings
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Task Queue
celery[redis]==5.4.0
redis==5.0.4

# Storage
boto3==1.34.100

# HTTP
httpx==0.27.0
beautifulsoup4==4.12.3

# AI
openai==1.30.1
anthropic==0.28.0
numpy==2.1.1
scikit-learn==1.5.1

# Google OAuth / GSC
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.129.0

# Utilities
python-dotenv==1.0.1
email-validator==2.1.1

# Rate Limiting & Observability
slowapi==0.1.9
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.47.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.27.0
